Split basis in several files.
