+++ skfda/_utils/__init__.py
@@ -2,4 +2,5 @@ from . import constants
 
 from ._utils import (_list_of_arrays, _coordinate_list,
                      _check_estimator, parameter_aliases,
-                     _to_grid, check_is_univariate)
+                     _to_grid, check_is_univariate,
+                     _same_domain)
