+++ skfda/representation/basis/__init__.py
@@ -1,0 +1,7 @@
+from ._basis import Basis
+from ._bspline import BSpline
+from ._coefficients_transformer import CoefficientsTransformer
+from ._constant import Constant
+from ._fdatabasis import FDataBasis, FDataBasisDType
+from ._fourier import Fourier
+from ._monomial import Monomial
