+++ skfda/representation/basis/_bspline.py
@@ -1,0 +1,453 @@
+from numpy import polyder, polyint, polymul, polyval
+from scipy.interpolate import BSpline as SciBSpline
+from scipy.interpolate import PPoly
+import scipy.interpolate
+
+import numpy as np
+
+from ..._utils import _list_of_arrays
+from ..._utils import _same_domain
+from ._basis import Basis
+
+
+class BSpline(Basis):
+    r"""BSpline basis.
+
+    BSpline basis elements are defined recursively as:
+
+    .. math::
+        B_{i, 1}(x) = 1 \quad \text{if } t_i \le x < t_{i+1},
+        \quad 0 \text{ otherwise}
+
+    .. math::
+        B_{i, k}(x) = \frac{x - t_i}{t_{i+k} - t_i} B_{i, k-1}(x)
+        + \frac{t_{i+k+1} - x}{t_{i+k+1} - t_{i+1}} B_{i+1, k-1}(x)
+
+    Where k indicates the order of the spline.
+
+    Implementation details: In order to allow a discontinuous behaviour at
+    the boundaries of the domain it is necessary to placing m knots at the
+    boundaries [RS05]_. This is automatically done so that the user only has to
+    specify a single knot at the boundaries.
+
+    Attributes:
+        domain_range (tuple): A tuple of length 2 containing the initial and
+            end values of the interval over which the basis can be evaluated.
+        n_basis (int): Number of functions in the basis.
+        order (int): Order of the splines. One greather than their degree.
+        knots (list): List of knots of the spline functions.
+
+    Examples:
+        Constructs specifying number of basis and order.
+
+        >>> bss = BSpline(n_basis=8, order=4)
+
+        If no order is specified defaults to 4 because cubic splines are
+        the most used. So the previous example is the same as:
+
+        >>> bss = BSpline(n_basis=8)
+
+        It is also possible to create a BSpline basis specifying the knots.
+
+        >>> bss = BSpline(knots=[0, 0.2, 0.4, 0.6, 0.8, 1])
+
+        Once we create a basis we can evaluate each of its functions at a
+        set of points.
+
+        >>> bss = BSpline(n_basis=3, order=3)
+        >>> bss.evaluate([0, 0.5, 1])
+        array([[ 1.  ,  0.25,  0.  ],
+               [ 0.  ,  0.5 ,  0.  ],
+               [ 0.  ,  0.25,  1.  ]])
+
+        And evaluates first derivative
+
+        >>> bss.evaluate([0, 0.5, 1], derivative=1)
+        array([[-2., -1.,  0.],
+               [ 2.,  0., -2.],
+               [ 0.,  1.,  2.]])
+
+    References:
+        .. [RS05] Ramsay, J., Silverman, B. W. (2005). *Functional Data
+            Analysis*. Springer. 50-51.
+
+    """
+
+    def __init__(self, domain_range=None, n_basis=None, order=4, knots=None):
+        """Bspline basis constructor.
+
+        Args:
+            domain_range (tuple, optional): Definition of the interval where
+                the basis defines a space. Defaults to (0,1) if knots are not
+                specified. If knots are specified defaults to the first and
+                last element of the knots.
+            n_basis (int, optional): Number of splines that form the basis.
+            order (int, optional): Order of the splines. One greater that
+                their degree. Defaults to 4 which mean cubic splines.
+            knots (array_like): List of knots of the splines. If domain_range
+                is specified the first and last elements of the knots have to
+                match with it.
+
+        """
+
+        if domain_range is not None:
+            domain_range = _list_of_arrays(domain_range)
+
+            if len(domain_range) != 1:
+                raise ValueError("Domain range should be unidimensional.")
+
+            domain_range = domain_range[0]
+
+        # Knots default to equally space points in the domain_range
+        if knots is None:
+            if n_basis is None:
+                raise ValueError("Must provide either a list of knots or the"
+                                 "number of basis.")
+        else:
+            knots = list(knots)
+            knots.sort()
+            if domain_range is None:
+                domain_range = (knots[0], knots[-1])
+            else:
+                if domain_range[0] != knots[0] or domain_range[1] != knots[-1]:
+                    raise ValueError("The ends of the knots must be the same "
+                                     "as the domain_range.")
+
+        # n_basis default to number of knots + order of the splines - 2
+        if n_basis is None:
+            n_basis = len(knots) + order - 2
+
+        if (n_basis - order + 2) < 2:
+            raise ValueError(f"The number of basis ({n_basis}) minus the "
+                             f"order of the bspline ({order}) should be "
+                             f"greater than 3.")
+
+        self.order = order
+        self.knots = None if knots is None else list(knots)
+        super().__init__(domain_range, n_basis)
+
+        # Checks
+        if self.n_basis != self.order + len(self.knots) - 2:
+            raise ValueError(f"The number of basis ({self.n_basis}) has to "
+                             f"equal the order ({self.order}) plus the "
+                             f"number of knots ({len(self.knots)}) minus 2.")
+
+    @property
+    def knots(self):
+        if self._knots is None:
+            return list(np.linspace(*self.domain_range[0],
+                                    self.n_basis - self.order + 2))
+        else:
+            return self._knots
+
+    @knots.setter
+    def knots(self, value):
+        self._knots = value
+
+    def _evaluation_knots(self):
+        """
+        Get the knots adding m knots to the boundary in order to allow a
+        discontinuous behaviour at the boundaries of the domain [RS05]_.
+
+        References:
+            .. [RS05] Ramsay, J., Silverman, B. W. (2005). *Functional Data
+                Analysis*. Springer. 50-51.
+        """
+        return np.array([self.knots[0]] * (self.order - 1) + self.knots +
+                        [self.knots[-1]] * (self.order - 1))
+
+    def _evaluate(self, eval_points, derivative=0):
+        """Compute the basis or its derivatives given a list of values.
+
+        It uses the scipy implementation of BSplines to compute the values
+        for each element of the basis.
+
+        Args:
+            eval_points (array_like): List of points where the basis system is
+                evaluated.
+            derivative (int, optional): Order of the derivative. Defaults to 0.
+
+        Returns:
+            (:obj:`numpy.darray`): Matrix whose rows are the values of the each
+            basis function or its derivatives at the values specified in
+            eval_points.
+
+        Implementation details: In order to allow a discontinuous behaviour at
+        the boundaries of the domain it is necessary to placing m knots at the
+        boundaries [RS05]_. This is automatically done so that the user only
+        has to specify a single knot at the boundaries.
+
+        References:
+            .. [RS05] Ramsay, J., Silverman, B. W. (2005). *Functional Data
+                Analysis*. Springer. 50-51.
+
+        """
+        if derivative > (self.order - 1):
+            return np.zeros((self.n_basis, len(eval_points)))
+
+        # Places m knots at the boundaries
+        knots = self._evaluation_knots()
+
+        # c is used the select which spline the function splev below computes
+        c = np.zeros(len(knots))
+
+        # Initialise empty matrix
+        mat = np.empty((self.n_basis, len(eval_points)))
+
+        # For each basis computes its value for each evaluation point
+        for i in range(self.n_basis):
+            # write a 1 in c in the position of the spline calculated in each
+            # iteration
+            c[i] = 1
+            # compute the spline
+            mat[i] = scipy.interpolate.splev(eval_points, (knots, c,
+                                                           self.order - 1),
+                                             der=derivative)
+            c[i] = 0
+
+        return mat
+
+    def _derivative(self, coefs, order=1):
+        deriv_splines = [self._to_scipy_BSpline(coefs[i]).derivative(order)
+                         for i in range(coefs.shape[0])]
+
+        deriv_coefs = [BSpline._from_scipy_BSpline(spline)[1]
+                       for spline in deriv_splines]
+
+        deriv_basis = BSpline._from_scipy_BSpline(deriv_splines[0])[0]
+
+        return deriv_basis, np.array(deriv_coefs)[:, 0:deriv_basis.n_basis]
+
+    def _penalty(self, lfd):
+
+        coefs = lfd.constant_weights()
+        if coefs is None:
+            return NotImplemented
+
+        nonzero = np.flatnonzero(coefs)
+
+        # All derivatives above the order of the spline are effectively
+        # zero
+        nonzero = nonzero[nonzero < self.order]
+
+        if len(nonzero) == 0:
+            return np.zeros((self.n_basis, self.n_basis))
+
+        # We will only deal with one nonzero coefficient right now
+        if len(nonzero) != 1:
+            return NotImplemented
+
+        derivative_degree = nonzero[0]
+
+        if derivative_degree == self.order - 1:
+            # The derivative of the bsplines are constant in the intervals
+            # defined between knots
+            knots = np.array(self.knots)
+            mid_inter = (knots[1:] + knots[:-1]) / 2
+            constants = self.evaluate(mid_inter,
+                                      derivative=derivative_degree).T
+            knots_intervals = np.diff(self.knots)
+            # Integration of product of constants
+            return constants.T @ np.diag(knots_intervals) @ constants
+
+        # We only deal with the case without zero length intervals
+        # for now
+        if np.any(np.diff(self.knots) == 0):
+            return NotImplemented
+
+        # Compute exactly using the piecewise polynomial
+        # representation of splines
+
+        # Places m knots at the boundaries
+        knots = self._evaluation_knots()
+
+        # c is used the select which spline the function
+        # PPoly.from_spline below computes
+        c = np.zeros(len(knots))
+
+        # Initialise empty list to store the piecewise polynomials
+        ppoly_lst = []
+
+        no_0_intervals = np.where(np.diff(knots) > 0)[0]
+
+        # For each basis gets its piecewise polynomial representation
+        for i in range(self.n_basis):
+
+            # Write a 1 in c in the position of the spline
+            # transformed in each iteration
+            c[i] = 1
+
+            # Gets the piecewise polynomial representation and gets
+            # only the positions for no zero length intervals
+            # This polynomial are defined relatively to the knots
+            # meaning that the column i corresponds to the ith knot.
+            # Let the ith knot be a
+            # Then f(x) = pp(x - a)
+            pp = PPoly.from_spline((knots, c, self.order - 1))
+            pp_coefs = pp.c[:, no_0_intervals]
+
+            # We have the coefficients for each interval in coordinates
+            # (x - a), so we will need to subtract a when computing the
+            # definite integral
+            ppoly_lst.append(pp_coefs)
+            c[i] = 0
+
+        # Now for each pair of basis computes the inner product after
+        # applying the linear differential operator
+        penalty_matrix = np.zeros((self.n_basis, self.n_basis))
+        for interval in range(len(no_0_intervals)):
+            for i in range(self.n_basis):
+                poly_i = np.trim_zeros(ppoly_lst[i][:,
+                                                    interval], 'f')
+                if len(poly_i) <= derivative_degree:
+                    # if the order of the polynomial is lesser or
+                    # equal to the derivative the result of the
+                    # integral will be 0
+                    continue
+                # indefinite integral
+                derivative = polyder(poly_i, derivative_degree)
+                square = polymul(derivative, derivative)
+                integral = polyint(square)
+
+                # definite integral
+                penalty_matrix[i, i] += np.diff(polyval(
+                    integral, self.knots[interval: interval + 2]
+                    - self.knots[interval]))[0]
+
+                for j in range(i + 1, self.n_basis):
+                    poly_j = np.trim_zeros(ppoly_lst[j][:,
+                                                        interval], 'f')
+                    if len(poly_j) <= derivative_degree:
+                        # if the order of the polynomial is lesser
+                        # or equal to the derivative the result of
+                        # the integral will be 0
+                        continue
+                        # indefinite integral
+                    integral = polyint(
+                        polymul(polyder(poly_i, derivative_degree),
+                                polyder(poly_j, derivative_degree)))
+                    # definite integral
+                    penalty_matrix[i, j] += np.diff(polyval(
+                        integral, self.knots[interval: interval + 2]
+                        - self.knots[interval])
+                    )[0]
+                    penalty_matrix[j, i] = penalty_matrix[i, j]
+        return penalty_matrix
+
+    def rescale(self, domain_range=None):
+        r"""Return a copy of the basis with a new domain range, with the
+            corresponding values rescaled to the new bounds.
+            The knots of the BSpline will be rescaled in the new interval.
+
+            Args:
+                domain_range (tuple, optional): Definition of the interval
+                    where the basis defines a space. Defaults uses the same as
+                    the original basis.
+        """
+
+        knots = np.array(self.knots, dtype=np.dtype('float'))
+
+        if domain_range is not None:  # Rescales the knots
+            knots -= knots[0]
+            knots *= ((domain_range[1] - domain_range[0]
+                       ) / (self.knots[-1] - self.knots[0]))
+            knots += domain_range[0]
+
+            # Fix possible round error
+            knots[0] = domain_range[0]
+            knots[-1] = domain_range[1]
+
+        else:
+            # TODO: Allow multiple dimensions
+            domain_range = self.domain_range[0]
+
+        return BSpline(domain_range, self.n_basis, self.order, knots)
+
+    def __repr__(self):
+        """Representation of a BSpline basis."""
+        return (f"{self.__class__.__name__}(domain_range={self.domain_range}, "
+                f"n_basis={self.n_basis}, order={self.order}, "
+                f"knots={self.knots})")
+
+    def __eq__(self, other):
+        """Equality of Basis"""
+        return (super().__eq__(other)
+                and self.order == other.order
+                and self.knots == other.knots)
+
+    def basis_of_product(self, other):
+        from ._constant import Constant
+
+        """Multiplication of two Bspline Basis"""
+        if not _same_domain(self, other):
+            raise ValueError("Ranges are not equal.")
+
+        if isinstance(other, Constant):
+            return other.rbasis_of_product(self)
+
+        if isinstance(other, BSpline):
+            uniqueknots = np.union1d(self.inknots, other.inknots)
+
+            multunique = np.zeros(len(uniqueknots), dtype=np.int32)
+            for i in range(len(uniqueknots)):
+                mult1 = np.count_nonzero(self.inknots == uniqueknots[i])
+                mult2 = np.count_nonzero(other.inknots == uniqueknots[i])
+                multunique[i] = max(mult1, mult2)
+
+            m2 = 0
+            allknots = np.zeros(np.sum(multunique))
+            for i in range(len(uniqueknots)):
+                m1 = m2
+                m2 = m2 + multunique[i]
+                allknots[m1:m2] = uniqueknots[i]
+
+            norder1 = self.n_basis - len(self.inknots)
+            norder2 = other.n_basis - len(other.inknots)
+            norder = min(norder1 + norder2 - 1, 20)
+
+            allbreaks = ([self.domain_range[0][0]] +
+                         np.ndarray.tolist(allknots) +
+                         [self.domain_range[0][1]])
+            n_basis = len(allbreaks) + norder - 2
+            return BSpline(self.domain_range, n_basis, norder, allbreaks)
+        else:
+            norder = min(self.n_basis - len(self.inknots) + 2, 8)
+            n_basis = max(self.n_basis + other.n_basis, norder + 1)
+            return BSpline(self.domain_range, n_basis, norder)
+
+    def rbasis_of_product(self, other):
+        """Multiplication of a Bspline Basis with other basis"""
+
+        norder = min(self.n_basis - len(self.inknots) + 2, 8)
+        n_basis = max(self.n_basis + other.n_basis, norder + 1)
+        return BSpline(self.domain_range, n_basis, norder)
+
+    def _to_R(self):
+        drange = self.domain_range[0]
+        return ("create.bspline.basis(rangeval = c(" + str(drange[0]) + "," +
+                str(drange[1]) + "), nbasis = " + str(self.n_basis) +
+                ", norder = " + str(self.order) + ", breaks = " +
+                self._list_to_R(self.knots) + ")")
+
+    def _to_scipy_BSpline(self, coefs):
+
+        knots = np.concatenate((
+            np.repeat(self.knots[0], self.order - 1),
+            self.knots,
+            np.repeat(self.knots[-1], self.order - 1)))
+
+        return SciBSpline(knots, coefs, self.order - 1)
+
+    @staticmethod
+    def _from_scipy_BSpline(bspline):
+        order = bspline.k
+        knots = bspline.t[order: -order]
+        coefs = bspline.c
+        domain_range = [knots[0], knots[-1]]
+
+        return BSpline(domain_range, order=order + 1, knots=knots), coefs
+
+    @property
+    def inknots(self):
+        """Return number of basis."""
+        return self.knots[1:len(self.knots) - 1]
